Added new classes and refactored GameContext
